Rename spec file
